backend/requirements.txt
# FastAPI & Web Server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0
alembic>=1.13.0

# Task Queue
celery[redis]>=5.3.0
redis>=5.0.0

# HTTP Client (async)
httpx>=0.26.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Validation & Settings
pydantic>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.1.0

# Notifications
sendgrid>=6.11.0
twilio>=8.10.0

# Payments
stripe>=7.0.0

# AWS S3
boto3>=1.34.0

# Utilities
python-dateutil>=2.8.2

# Development
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
